Better support for custom prefixes
Custom prefixes can be specified on a per-message basis just like before, but now it's done by passing a function to a method rather than overriding a method. Additionally, the function gets an incomplete context object (missing commandName and prefix, duh) as an argument, so it can take advantage of context additions set via extendContext.